requirements.txt
tensorflow
pytorch
matplotlib
nump
